ADD: Transactions paginated and cache

+++ backend/Application/Shared/PaginatedList.cs
@@ -2,7 +2,7 @@ namespace Application.Shared;
 
 public class PaginatedList<T>(
     List<T> items,
-    int count,
+    long count,
     int pageNumber,
     int pageSize
 )
@@ -10,6 +10,6 @@ public class PaginatedList<T>(
     public List<T> Items { get; set; } = items;
     public int PageNumber { get; set; } = pageNumber;
     public int PageSize { get; set; } = pageSize;
-    public int TotalCount { get; set; } = count;
+    public long TotalCount { get; set; } = count;
     public int TotalPages => (int)Math.Ceiling(TotalCount / (double)PageSize);
 }
